+++ src/Addons/34thPRC_ArmourStandard/data/config_weapons.hpp
@@ -1,5 +1,6 @@
 #include "beret\config_weapons.hpp"
 #include "marines\config_weapons.hpp"
 #include "mkiv\config_weapons.hpp"
+#include "mkvb\config_weapons.hpp"
 #include "odst\config_weapons.hpp"
 #include "spi\config_weapons.hpp"