+++ src/Addons/34thPRC_ArmourStandard/data/mkvb/config_weapons.hpp
@@ -1,0 +1,44 @@
+class MA_Mjolnir_MKVB_Helmet; // MA_Armor
+class MA_Mjolnir_Scout_Helmet; // MA_Armor
+class MA_Mjolnir_MKV_Helmet; // MA_Armor
+class 34thPRC_ArmourStandard_MKVB_Helmet : MA_Mjolnir_MKVB_Helmet
+{
+    displayName="[34th] MK V[B] Helmet";
+    author= "Vazya";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourStandard\data\mkvb\MKVB\Helmet\34_MKVB_Helmet.paa",
+        "34thPRC_ArmourStandard\data\mkvb\MKVB\Visors\34_MKVB_Helmet_Visor_Gold.paa"
+    };
+};
+class 34thPRC_ArmourStandard_MKVB_Scout_Helmet : MA_Mjolnir_Scout_Helmet
+{
+    displayName="[34th] Scout Helmet";
+    author= "Vazya";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourStandard\data\mkvb\Scout\Helmet\34_Scout_Helmet.paa",
+        "34thPRC_ArmourStandard\data\mkvb\Scout\Visors\34_Scout_Helmet_Visor_Gold.paa"
+    };
+};
+class 34thPRC_ArmourStandard_MKVB_MKV_Helmet : MA_Mjolnir_MKV_Helmet
+{
+    displayName="[34th] MK V Helmet";
+    author= "Vazya";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourStandard\data\mkvb\MKV\Helmet\34_MKV_Helmet.paa",
+        "34thPRC_ArmourStandard\data\mkvb\MKV\Visors\34_MKV_Helmet_Visor_Gold.paa"
+    };
+};
+class MA_MKVB_Armor; // MA_Armor
+class 34thPRC_ArmourStandard_MKVB_Armor: MA_MKVB_Armor
+{
+    displayName="[34th] MK V[B] Armor";
+    author= "Vazya"
+	hiddenSelectionsTextures[]=
+    {
+            "34thPRC_ArmourStandard\data\mkvb\MKVB\34_MKVB_Upper_CO.paa",    //camo4 (Upper Armor)
+            "34thPRC_ArmourStandard\data\mkvb\MKVB\34_MKVB_Lower_CO.paa",    //camo5 (Lower Armor)
+    };
+};